Honor appConfig title as it was entered
If I specifically enter a title for the `appConfig`, I want it to be used as given. For example, I want a title of "HealtheIntent Package" but it will convert it to "Healthe Intent Package". The [default appConfig.title](https://github.com/cerner/terra-dev-site/blob/main/config/site/site.config.js#L57) already does the name casing so there is no need for this to be here.